test function
separate test function from virtual kubelet sources
